feat: rename /admin/form endpoint to /admin/forms, correct dir structure of /api/v3 routes (#1425)
