--- src/app/routes/admin/admin.routes.ts
@@ -1,7 +1,0 @@
-import { Router } from 'express'
-
-import { AdminFormRouter } from './form/admin-form.routes'
-
-export const AdminRouter = Router()
-
-AdminRouter.use('/form', AdminFormRouter)
