--- src/app/routes/admin/form/index.ts
@@ -1,1 +1,0 @@
-export { AdminFormRouter } from './admin-form.routes'
