+++ src/app/routes/api/index.ts
@@ -1,0 +1,1 @@
+export { ApiRouter } from './api.routes'
