+++ src/app/routes/api/v3/admin/admin.routes.ts
@@ -1,0 +1,7 @@
+import { Router } from 'express'
+
+import { AdminFormsRouter } from './forms'
+
+export const AdminRouter = Router()
+
+AdminRouter.use('/forms', AdminFormsRouter)
