+++ src/app/routes/api/v3/admin/forms/index.ts
@@ -1,0 +1,1 @@
+export { AdminFormsRouter } from './admin-forms.routes'
