+++ src/app/routes/api/v3/admin/index.ts
