+++ src/app/routes/api/v3/index.ts
@@ -1,0 +1,1 @@
+export { V3Router } from './v3.routes'
