+++ src/app/routes/api/v3/v3.routes.ts
@@ -1,0 +1,7 @@
+import { Router } from 'express'
+
+import { AdminRouter } from './admin'
+
+export const V3Router = Router()
+
+V3Router.use('/admin', AdminRouter)
