+++ src/loaders/express/index.ts
@@ -25,7 +25,7 @@ import { SubmissionRouter } from '../../app/modules/submission/submission.routes
 import UserRouter from '../../app/modules/user/user.routes'
 import { VfnRouter } from '../../app/modules/verification/verification.routes'
 import apiRoutes from '../../app/routes'
-import { API_V3_ROOT_ROUTE, ApiRouter } from '../../app/routes/api.routes'
+import { ApiRouter } from '../../app/routes/api'
 import * as IntranetMiddleware from '../../app/services/intranet/intranet.middleware'
 import config from '../../config/config'
 
@@ -165,7 +165,7 @@ const loadExpressApp = async (connection: Connection) => {
   app.use(MYINFO_ROUTER_PREFIX, MyInfoRouter)
 
   // New routes in preparation for API refactor.
-  app.use(API_V3_ROOT_ROUTE, ApiRouter)
+  app.use('/api', ApiRouter)
 
   app.use(sentryMiddlewares())
